+++ read.go
@@ -16,7 +16,7 @@ package ice
 
 import "encoding/binary"
 
-func (s *Segment) getDocStoredMetaAndCompressed(docNum int) (meta, data []byte, err error) {
+func (s *Segment) getDocStoredMetaAndCompressed(docNum uint64) (meta, data []byte, err error) {
 	_, storedOffset, n, metaLen, dataLen, err := s.getDocStoredOffsets(docNum)
 	if err != nil {
 		return nil, nil, err
@@ -34,9 +34,9 @@ func (s *Segment) getDocStoredMetaAndCompressed(docNum int) (meta, data []byte, 
 	return meta, data, nil
 }
 
-func (s *Segment) getDocStoredOffsets(docNum int) (
+func (s *Segment) getDocStoredOffsets(docNum uint64) (
 	indexOffset, storedOffset, n, metaLen, dataLen uint64, err error) {
-	indexOffset = s.footer.storedIndexOffset + (fileAddrWidth * uint64(docNum))
+	indexOffset = s.footer.storedIndexOffset + (fileAddrWidth * docNum)
 
 	storedOffsetData, err := s.data.Read(int(indexOffset), int(indexOffset+fileAddrWidth))
 	if err != nil {
